+++ src/server/game/Server/Packets/ItemPackets.cpp
@@ -257,7 +257,7 @@ WorldPacket const* WorldPackets::Item::ItemPushResult::Write()
     _worldPacket << int32(DungeonEncounterID);
     _worldPacket << int32(BattlePetSpeciesID);
     _worldPacket << int32(BattlePetBreedID);
-    _worldPacket << uint32(BattlePetBreedQuality);
+    _worldPacket << uint8(BattlePetBreedQuality);
     _worldPacket << int32(BattlePetLevel);
     _worldPacket << ItemGUID;
