+++ src/server/game/Server/Packets/ItemPackets.h
@@ -349,7 +349,7 @@ namespace WorldPackets
             int32 DungeonEncounterID        = 0;
             int32 BattlePetSpeciesID        = 0;
             int32 BattlePetBreedID          = 0;
-            uint32 BattlePetBreedQuality    = 0;
+            uint8 BattlePetBreedQuality     = 0;
             int32 BattlePetLevel            = 0;
             ObjectGuid ItemGUID;
             bool Pushed                     = false;
